Introduced a new class which builds cachable metadata about a git repository
This makes the cache non-static and hopefully easier to maintain in the long wrong
