+++ src/GitVersionCore/BranchConfigurationCalculator.cs
@@ -12,9 +12,9 @@ namespace GitVersion
         /// <summary>
         /// Gets the <see cref="BranchConfig"/> for the current commit.
         /// </summary>
-        public static BranchConfig GetBranchConfiguration(Commit currentCommit, IRepository repository, bool onlyEvaluateTrackedBranches, Config config, Branch currentBranch, IList<Branch> excludedInheritBranches = null)
+        public static BranchConfig GetBranchConfiguration(GitVersionContext context, Branch targetBranch, IList<Branch> excludedInheritBranches = null)
         {
-            var matchingBranches = LookupBranchConfiguration(config, currentBranch).ToArray();
+            var matchingBranches = LookupBranchConfiguration(context.FullConfiguration, targetBranch).ToArray();
 
             BranchConfig branchConfiguration;
             if (matchingBranches.Length > 0)
@@ -25,7 +25,7 @@ namespace GitVersion
                 {
                     Logger.WriteWarning(string.Format(
                         "Multiple branch configurations match the current branch branchName of '{0}'. Using the first matching configuration, '{1}'. Matching configurations include: '{2}'",
-                        currentBranch.FriendlyName,
+                        targetBranch.FriendlyName,
                         branchConfiguration.Name,
                         string.Join("', '", matchingBranches.Select(b => b.Name))));
                 }
@@ -34,14 +34,14 @@ namespace GitVersion
             {
                 Logger.WriteInfo(string.Format(
                     "No branch configuration found for branch {0}, falling back to default configuration",
-                    currentBranch.FriendlyName));
+                    targetBranch.FriendlyName));
 
                 branchConfiguration = new BranchConfig { Name = string.Empty };
-                ConfigurationProvider.ApplyBranchDefaults(config, branchConfiguration, "");
+                ConfigurationProvider.ApplyBranchDefaults(context.FullConfiguration, branchConfiguration, "");
             }
 
             return branchConfiguration.Increment == IncrementStrategy.Inherit ?
-                InheritBranchConfiguration(onlyEvaluateTrackedBranches, repository, currentCommit, currentBranch, branchConfiguration, config, excludedInheritBranches) :
+                InheritBranchConfiguration(context, targetBranch, branchConfiguration, excludedInheritBranches) :
                 branchConfiguration;
         }
 
@@ -60,16 +60,18 @@ namespace GitVersion
             return config.Branches.Where(b => Regex.IsMatch(currentBranch.FriendlyName, "^" + b.Value.Regex, RegexOptions.IgnoreCase)).Select(kvp => kvp.Value);
         }
 
-        static BranchConfig InheritBranchConfiguration(bool onlyEvaluateTrackedBranches, IRepository repository, Commit currentCommit, Branch currentBranch, BranchConfig branchConfiguration, Config config, IList<Branch> excludedInheritBranches)
+        static BranchConfig InheritBranchConfiguration(GitVersionContext context, Branch targetBranch, BranchConfig branchConfiguration, IList<Branch> excludedInheritBranches)
         {
+            var repository = context.Repository;
+            var config = context.FullConfiguration;
             using (Logger.IndentLog("Attempting to inherit branch configuration from parent branch"))
             {
-                var excludedBranches = new[] { currentBranch };
+                var excludedBranches = new[] { targetBranch };
                 // Check if we are a merge commit. If so likely we are a pull request
-                var parentCount = currentCommit.Parents.Count();
+                var parentCount = context.CurrentCommit.Parents.Count();
                 if (parentCount == 2)
                 {
-                    excludedBranches = CalculateWhenMultipleParents(repository, currentCommit, ref currentBranch, excludedBranches);
+                    excludedBranches = CalculateWhenMultipleParents(repository, context.CurrentCommit, ref targetBranch, excludedBranches);
                 }
 
                 if (excludedInheritBranches == null)
@@ -90,11 +92,12 @@ namespace GitVersion
                 }
                 var branchesToEvaluate = repository.Branches.Except(excludedInheritBranches).ToList();
 
-                var branchPoint = currentBranch.FindCommitBranchWasBranchedFrom(repository, excludedInheritBranches.ToArray());
+                var branchPoint = context.RepostioryMetadataProvider
+                    .FindCommitBranchWasBranchedFrom(targetBranch, repository, excludedInheritBranches.ToArray());
                 List<Branch> possibleParents;
                 if (branchPoint == BranchCommit.Empty)
                 {
-                    possibleParents = currentCommit.GetBranchesContainingCommit(repository, branchesToEvaluate, true)
+                    possibleParents = context.RepostioryMetadataProvider.GetBranchesContainingCommit(context.CurrentCommit, repository, branchesToEvaluate, true)
                         // It fails to inherit Increment branch configuration if more than 1 parent;
                         // therefore no point to get more than 2 parents
                         .Take(2)
@@ -102,10 +105,12 @@ namespace GitVersion
                 }
                 else
                 {
-                    var branches = branchPoint.Commit.GetBranchesContainingCommit(repository, branchesToEvaluate, true).ToList();
+                    var branches = context.RepostioryMetadataProvider
+                        .GetBranchesContainingCommit(branchPoint.Commit, repository, branchesToEvaluate, true).ToList();
                     if (branches.Count > 1)
                     {
-                        var currentTipBranches = currentCommit.GetBranchesContainingCommit(repository, branchesToEvaluate, true).ToList();
+                        var currentTipBranches = context.RepostioryMetadataProvider
+                            .GetBranchesContainingCommit(context.CurrentCommit, repository, branchesToEvaluate, true).ToList();
                         possibleParents = branches.Except(currentTipBranches).ToList();
                     }
                     else
@@ -118,7 +123,7 @@ namespace GitVersion
 
                 if (possibleParents.Count == 1)
                 {
-                    var branchConfig = GetBranchConfiguration(currentCommit, repository, onlyEvaluateTrackedBranches, config, possibleParents[0], excludedInheritBranches);
+                    var branchConfig = GetBranchConfiguration(context, possibleParents[0], excludedInheritBranches);
                     return new BranchConfig(branchConfiguration)
                     {
                         Increment = branchConfig.Increment,
@@ -149,7 +154,7 @@ namespace GitVersion
                 Logger.WriteWarning(errorMessage + Environment.NewLine + Environment.NewLine + "Falling back to " + branchName + " branch config");
 
                 // To prevent infinite loops, make sure that a new branch was chosen.
-                if (LibGitExtensions.IsSameBranch(currentBranch, chosenBranch))
+                if (LibGitExtensions.IsSameBranch(targetBranch, chosenBranch))
                 {
                     Logger.WriteWarning("Fallback branch wants to inherit Increment branch configuration from itself. Using patch increment instead.");
                     return new BranchConfig(branchConfiguration)
@@ -158,7 +163,7 @@ namespace GitVersion
                     };
                 }
 
-                var inheritingBranchConfig = GetBranchConfiguration(currentCommit, repository, onlyEvaluateTrackedBranches, config, chosenBranch, excludedInheritBranches);
+                var inheritingBranchConfig = GetBranchConfiguration(context, chosenBranch, excludedInheritBranches);
                 return new BranchConfig(branchConfiguration)
                 {
                     Increment = inheritingBranchConfig.Increment,
